backend/requirements.txt
Django>=3.0.8,<3.1.0
psycopg2>=2.8.5,<2.9.0
flake8>=3.8.3,<3.9.0
djangorestframework>=3.11.0,<3.12.0
django-cors-headers>=3.4.0,<3.5.0
